[A64/Emitter] Fix code size assert cast
- Compare code_size.total to size_t offset()

+++ src/xenia/cpu/backend/a64/a64_emitter.cc
@@ -189,7 +189,7 @@ void* A64Emitter::Emplace(const EmitFunctionInfo& func_info,
   void* new_execute_address;
   void* new_write_address;
 
-  assert_true(func_info.code_size.total == offset());
+  assert_true(func_info.code_size.total == static_cast<size_t>(offset()));
 
   if (function) {
     code_cache_->PlaceGuestCode(function->address(), assembly_buffer.data(),
